update submitInContest() to handle user ranking data

+++ src/main/java/com/xjudge/entity/UserContest.java
@@ -38,7 +38,7 @@ public class UserContest extends BaseEntity<UserContestKey> {
 
     private Boolean isParticipant;
 
-    private Integer userContestPenalty;
+    private Long userContestPenalty;
 
     private Integer userContestRank;
